requirements.txt
# Python dependencies for Distributed Internet Topology Mapper

# Core dependencies
asyncio-mqtt>=0.16.1
pyyaml>=6.0

# IPFS integration
ipfshttpclient>=0.8.0a2

# Network tools
scapy>=2.5.0

# Bandwidth testing
speedtest-cli>=2.1.3
iperf3>=0.1.11

# API server
flask>=2.3.0
flask-cors>=4.0.0

# HTTP requests
requests>=2.31.0

# Data structures and utilities
python-dotenv>=1.0.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Build tools (for packaging)
pyinstaller>=6.0.0

# Code quality (development)
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
